+++ src/main/java/agents/electrolysis/ElectrolyzerStrategist.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.TreeMap;
 import agents.flexibility.DispatchSchedule;
 import agents.flexibility.Strategist;
+import agents.markets.meritOrder.sensitivities.PriceNoSensitivity;
 import de.dlr.gitlab.fame.agent.input.Make;
 import de.dlr.gitlab.fame.agent.input.ParameterData;
 import de.dlr.gitlab.fame.agent.input.ParameterData.MissingDataException;
@@ -25,15 +26,27 @@ public abstract class ElectrolyzerStrategist extends Strategist {
 		DISPATCH_FILE,
 		/** Schedules based on a moving target of hydrogen production totals and electricity prices */
 		SINGLE_AGENT_SIMPLE,
+		/** Schedules green hydrogen dispatch for monthly equivalence between electrolysis demand and renewable generation */
+		GREEN_HYDROGEN_MONTHLY
 	}
 
 	/** Planned production schedule for hydrogen in thermal MWh */
 	protected double[] scheduledChargedHydrogenTotal;
 	/** total actual hydrogen produced in thermal MWH */
 	protected double actualProducedHydrogen = 0;
-	/** the associated electrolysis unit */
+	/** Electricity price forecasts for planning horizon */
+	protected double[] electricityPriceForecasts;
+	/** Maximum willingness to pay for electricity compared to selling hydrogen */
+	protected double[] hydrogenSaleOpportunityCostsPerElectricMWH;
+	/** Electric power dedicated to electrolysis unit */
+	protected double[] electricDemandOfElectrolysisInMW;
+	/** Time stamps at the beginning of each scheduling period */
+	protected TimeStamp[] stepTimes;
+	/** The associated electrolysis unit */
 	protected Electrolyzer electrolyzer;
-	private DispatchSchedule schedule;
+	/** The Dispatch schedule controlled / created by this Strategist */
+	protected DispatchSchedule schedule;
+
 	private TreeMap<TimePeriod, Double> hydrogenPrices = new TreeMap<>();
 	private TimeSeries priceLimitOverrideInEURperMWH;
 
@@ -52,7 +65,15 @@ public abstract class ElectrolyzerStrategist extends Strategist {
 	protected ElectrolyzerStrategist(ParameterData input) throws MissingDataException {
 		super(input);
 		scheduledChargedHydrogenTotal = new double[scheduleDurationPeriods];
+		allocatePlanningArrays();
 		priceLimitOverrideInEURperMWH = input.getTimeSeriesOrDefault("PriceLimitOverrideInEURperMWH", null);
+	}
+
+	private void allocatePlanningArrays() {
+		electricityPriceForecasts = new double[forecastSteps];
+		hydrogenSaleOpportunityCostsPerElectricMWH = new double[forecastSteps];
+		electricDemandOfElectrolysisInMW = new double[forecastSteps];
+		stepTimes = new TimeStamp[forecastSteps];
 	}
 
 	/** Creates new electrolysis Strategist based on its associated input group
@@ -72,6 +93,9 @@ public abstract class ElectrolyzerStrategist extends Strategist {
 			case SINGLE_AGENT_SIMPLE:
 				strategist = new SingleAgentSimple(input, input.getGroup("Simple"));
 				break;
+			case GREEN_HYDROGEN_MONTHLY:
+				strategist = new MonthlyEquivalence(input);
+				break;
 			default:
 				throw new RuntimeException(ERR_UNKNOWN_STRATEGIST + type);
 		}
@@ -84,6 +108,38 @@ public abstract class ElectrolyzerStrategist extends Strategist {
 	 * @param electrolyzer to be assigned to this strategist */
 	protected final void setElectrolyzer(Electrolyzer electrolyzer) {
 		this.electrolyzer = electrolyzer;
+	}
+
+	/** copies forecasted prices from sensitivities
+	 * 
+	 * @param startTime first time period that update is done for */
+	protected void updateElectricityPriceForecasts(TimePeriod startTime) {
+		for (int period = 0; period < forecastSteps; period++) {
+			TimePeriod timePeriod = startTime.shiftByDuration(period);
+			PriceNoSensitivity priceObject = ((PriceNoSensitivity) getSensitivityForPeriod(timePeriod));
+			electricityPriceForecasts[period] = priceObject != null ? priceObject.getPriceForecast() : 0;
+		}
+	}
+
+	/** calculates electricity price equivalent of opportunity costs for selling hydrogen with expected prices
+	 * 
+	 * @param startTime first time period that update is done for */
+	protected void updateOpportunityCosts(TimePeriod startTime) {
+		for (int period = 0; period < forecastSteps; period++) {
+			TimePeriod timePeriod = startTime.shiftByDuration(period);
+			double hydrogenPriceInEURperThermalMWH = getHydrogenPriceForPeriod(timePeriod);
+			hydrogenSaleOpportunityCostsPerElectricMWH[period] = hydrogenPriceInEURperThermalMWH
+					* electrolyzer.getConversionFactor();
+		}
+	}
+
+	/** set start time of each hour in forecast interval
+	 * 
+	 * @param startTime first time period that update is done for */
+	protected void updateStepTimes(TimePeriod startTime) {
+		for (int hour = 0; hour < forecastSteps; hour++) {
+			stepTimes[hour] = startTime.shiftByDuration(hour).getStartTime();
+		}
 	}
 
 	/** Provides forecast for electricity demand at given time;<br>
@@ -148,5 +204,33 @@ public abstract class ElectrolyzerStrategist extends Strategist {
 	protected double getHydrogenPriceForPeriod(TimePeriod timePeriod) {
 		Double priceForecast = hydrogenPrices.get(timePeriod);
 		return priceForecast != null ? priceForecast : Double.MAX_VALUE;
+	}
+
+	/** @param endOfHorizon last step of planning horizon
+	 * @return Hour with highest economic potential among those with remaining production capabilities; -1 of no capability left in
+	 *         any hour */
+	protected int getHourWithHighestEconomicPotential(int endOfHorizon) {
+		int bestHour = -1;
+		double highestPotential = 0; // start with 0 to ignore hours with negative economic potential
+		for (int hour = 0; hour < endOfHorizon; hour++) {
+			double economicPotential = getEconomicHydrogenPotential(hour);
+			if (economicPotential > highestPotential && getRemainingPowerInMW(hour) > 0) {
+				highestPotential = economicPotential;
+				bestHour = hour;
+			}
+		}
+		return bestHour;
+	}
+
+	/** @param hour to evaluate
+	 * @return remaining (unused) electrolyzer electric capacity in MW for given hour */
+	protected double getRemainingPowerInMW(int hour) {
+		return Math.max(0., electrolyzer.getPeakPower(stepTimes[hour]) - electricDemandOfElectrolysisInMW[hour]);
+	}
+
+	/** @param hour to evaluate
+	 * @return Economic potential in given hour to purchase electricity and sell hydrogen */
+	protected double getEconomicHydrogenPotential(int hour) {
+		return hydrogenSaleOpportunityCostsPerElectricMWH[hour] - electricityPriceForecasts[hour];
 	}
 }