+++ src/main/java/agents/electrolysis/SingleAgentSimple.java
@@ -26,19 +26,18 @@ import util.Polynomial;
 public class SingleAgentSimple extends ElectrolyzerStrategist {
 	/** Inputs specific to {@link SingleAgentSimple} electrolyzer strategists */
 	public static final Tree parameters = Make.newTree().add(
-			Make.newSeries("HydrogenProductionTargetInMWH").help("How much hydrogen to produce per production interval"),
-			Make.newInt("ProductionTargetIntervalInHours").help("How many hours a production interval spans"),
-			Make.newDouble("PriceSensitivityFunction").list().help("Price change per additional load in EUR per MWH per MWH"),
-			Make.newDouble("PowerStepInMW")).buildTree();
+			Make.newSeries("HydrogenProductionTargetInMWH").optional()
+					.help("How much hydrogen to produce per production interval"),
+			Make.newInt("ProductionTargetIntervalInHours").optional().help("How many hours a production interval spans"),
+			Make.newDouble("PriceSensitivityFunction").optional().list()
+					.help("Price change per additional load in EUR per MWH per MWH"),
+			Make.newDouble("PowerStepInMW").optional()).buildTree();
 
 	private final TimeSeries productionTargets;
 	private final int productionInterval;
 	private final Polynomial priceSensitivity;
 	private final double powerStepInMW;
-	private double[] electricityPriceForecasts;
-	private double[] hydrogenSaleOpportunityCostsPerElectricMWH;
-	private double[] electricDemandOfElectrolysisInMW;
-	private TimeStamp[] stepTimes;
+
 	private TimePeriod productionPeriod;
 
 	/** Create new {@link SingleAgentSimple}
@@ -52,14 +51,6 @@ public class SingleAgentSimple extends ElectrolyzerStrategist {
 		productionInterval = specificInput.getInteger("ProductionTargetIntervalInHours");
 		priceSensitivity = new Polynomial(specificInput.getList("PriceSensitivityFunction", Double.class));
 		powerStepInMW = specificInput.getDouble("PowerStepInMW");
-		allocatePlanningResources();
-	}
-
-	private void allocatePlanningResources() {
-		electricityPriceForecasts = new double[forecastSteps];
-		electricDemandOfElectrolysisInMW = new double[forecastSteps];
-		stepTimes = new TimeStamp[forecastSteps];
-		hydrogenSaleOpportunityCostsPerElectricMWH = new double[forecastSteps];
 		Arrays.fill(priceScheduleInEURperMWH, Constants.SCARCITY_PRICE_IN_EUR_PER_MWH);
 	}
 
@@ -92,36 +83,10 @@ public class SingleAgentSimple extends ElectrolyzerStrategist {
 		}
 	}
 
-	/** copies forecasted prices from sensitivities */
-	private void updateElectricityPriceForecasts(TimePeriod startTime) {
-		for (int period = 0; period < forecastSteps; period++) {
-			TimePeriod timePeriod = startTime.shiftByDuration(period);
-			PriceNoSensitivity priceObject = ((PriceNoSensitivity) getSensitivityForPeriod(timePeriod));
-			electricityPriceForecasts[period] = priceObject != null ? priceObject.getPriceForecast() : 0;
-		}
-	}
-
-	/** calculates electricity price equivalent of opportunity costs for selling hydrogen with expected prices */
-	private void updateOpportunityCosts(TimePeriod startTime) {
-		for (int period = 0; period < forecastSteps; period++) {
-			TimePeriod timePeriod = startTime.shiftByDuration(period);
-			double hydrogenPriceInEURperThermalMWH = getHydrogenPriceForPeriod(timePeriod);
-			hydrogenSaleOpportunityCostsPerElectricMWH[period] = hydrogenPriceInEURperThermalMWH
-					* electrolyzer.getConversionFactor();
-		}
-	}
-
 	/** updates maximum bidding prices for electricity based on opportunity costs for hydrogen */
 	private void updateBiddingPriceLimits() {
 		for (int period = 0; period < scheduleDurationPeriods; period++) {
 			priceScheduleInEURperMWH[period] = hydrogenSaleOpportunityCostsPerElectricMWH[period];
-		}
-	}
-
-	/** set start time of each hour in forecast interval */
-	private void updateStepTimes(TimePeriod timePeriod) {
-		for (int hour = 0; hour < forecastSteps; hour++) {
-			stepTimes[hour] = timePeriod.shiftByDuration(hour).getStartTime();
 		}
 	}
 
@@ -165,7 +130,7 @@ public class SingleAgentSimple extends ElectrolyzerStrategist {
 	/** distribute conversion power such that costs are minimal assuming rising prices due to additional demand */
 	private void optimiseDispatch(double hydrogenProductionTargetInMWH) {
 		hourLoop: while (hydrogenProductionTargetInMWH > 0) {
-			int bestHour = getHourWithLowestPriceAndProductionCapability();
+			int bestHour = getHourWithHighestEconomicPotential(forecastSteps);
 			if (bestHour < 0) {
 				break hourLoop;
 			}
@@ -179,25 +144,6 @@ public class SingleAgentSimple extends ElectrolyzerStrategist {
 			electricityPriceForecasts[bestHour] += calcPriceIncrease(electricityPriceForecasts[bestHour],
 					electricEnergyInMWH);
 		}
-	}
-
-	/** @return Hour with lowest price among those with remaining production capabilities; -1 of no capability left in any hour */
-	private int getHourWithLowestPriceAndProductionCapability() {
-		int bestHour = -1;
-		double highestPotential = 0; // start with 0 to ignore hours with negative economic potential
-		for (int hour = 0; hour < forecastSteps; hour++) {
-			double economicPotential = hydrogenSaleOpportunityCostsPerElectricMWH[hour] - electricityPriceForecasts[hour];
-			if (economicPotential > highestPotential && getRemainingPowerInMW(hour) > 0) {
-				highestPotential = economicPotential;
-				bestHour = hour;
-			}
-		}
-		return bestHour;
-	}
-
-	/** @return remaining electric power in MW for given hour */
-	private double getRemainingPowerInMW(int hour) {
-		return Math.max(0., electrolyzer.getPeakPower(stepTimes[hour]) - electricDemandOfElectrolysisInMW[hour]);
 	}
 
 	/** @return estimate price increase due to additional demand */